file listener compatible with nc33-34
